Boilerplate and layout
This will be a video sharing app that allows users to paste a youtube link and watch the video together with any of their friends that they want to send the link to.

To start i have setup a new vite project and installed eslint, i have created the planning file and the file structure i will need for this.

I have also begun work on the homepage as it should be quite simple with a hero image and a brief introduction to the site and a cta button to create a room.

Once they have created a room they will be able to play and queue videos and send their friends the link so they can watch along.