Video Embed and Sync
I have implemented the logic for the synching of the video. After going over the youtube iframe api i found a better way to get my embeded video on screen. I have then created some helper functions that will listen for changes to my db state to tell the other listening clients to seek, play or pause the video to the needed place to match the changes made by one of the viewers. This works quite well albeit with a little delay between the videos.

I have introduced some bugs that also need fixing, namely that whenever a new user joins through a video that is already playing the video is restarted. I believe i know the way to fix it i just cant seem to make it work but the iframe api lets us specify a time for the video to start from. The issue here being that each new instance is a new mount thus instantly restarting the video for all users as the time is set back to 0.

I have also set up the backend with firebase to make the synchronization work. I have built out a fairly simple structure of having a rooms collection within which each room is stored under its id which is created when a user creates a room.

I have also added the functionality for creating rooms which will create a db entry and navigate the user accordingly