Web socket experimentation
I have never used web sockets but ive been researching realtime data solutions and found this could be a great option to get users watching a synchronized video.

I am very new to it however so as of right now my implementation does not work. I intend to make a much smaller scale app to play around with it a little and see if it would fit the scope of this project. I did also have a play around with a real time database in firebase, latency so far is not an issue using firestore however.